+++ opentelemetry-sdk/CHANGELOG.md
@@ -4,7 +4,24 @@
 
 - Add "metrics", "logs" to default features. With this, default feature list is
   "trace", "metrics" and "logs".
+- Add `with_resource` on Builder for LoggerProvider, replacing the `with_config`
+  method. Instead of using
+  `.with_config(Config::default().with_resource(RESOURCE::default()))` users
+  must now use `.with_resource(RESOURCE::default())` to configure Resource on
+  logger provider.
 - Removed dependency on `ordered-float`.
+
+- **Breaking** [1726](https://github.com/open-telemetry/opentelemetry-rust/pull/1726)
+  Update `LogProcessor::emit() method to take mutable reference to LogData. This is breaking
+  change for LogProcessor developers. If the processor needs to invoke the exporter 
+  asynchronously, it should clone the data to ensure it can be safely processed without 
+  lifetime issues. Any changes made to the log data before cloning in this method will be
+  reflected in the next log processor in the chain, as well as to the exporter.
+- **Breaking** [1726](https://github.com/open-telemetry/opentelemetry-rust/pull/1726)
+ Update `LogExporter::export() method to accept a batch of log data, which can be either a 
+ reference or owned `LogData`. If the exporter needs to process the log data
+ asynchronously, it should clone the log data to ensure it can be safely processed without 
+ lifetime issues. 
 
 ## v0.23.0